+++ src/main/java/janggi/piece/Piece.java
@@ -2,6 +2,8 @@ package janggi.piece;
 
 import janggi.board.Board;
 import janggi.position.Position;
+import janggi.rule.MoveVector;
+import janggi.rule.MovingRule;
 import janggi.rule.MovingRules;
 
 public sealed abstract class Piece permits Cannon, Chariot, Elephant, General, Guard, Horse, Soldier {
@@ -24,6 +26,22 @@ public sealed abstract class Piece permits Cannon, Chariot, Elephant, General, G
         return isValidDestination(end, board);
     }
 
+    protected boolean cannotMoveThrough(final Position start, final Position end, final Board board) {
+        final MovingRule matchRule = movingRules.findMatchRule(start, end);
+        Position route = start;
+        for (MoveVector vector : matchRule.getVectorsWithoutLast()) {
+            route = route.add(vector);
+            if (board.isPresent(route)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    protected boolean isValidDestination(final Position end, final Board board) {
+        return !board.isPresentSameTeam(team, end);
+    }
+
     public boolean isSameTeam(final Team team) {
         return this.team == team;
     }
@@ -31,10 +49,6 @@ public sealed abstract class Piece permits Cannon, Chariot, Elephant, General, G
     public Team getTeam() {
         return team;
     }
-
-    protected abstract boolean cannotMoveThrough(final Position start, final Position end, final Board board);
-
-    protected abstract boolean isValidDestination(final Position end, final Board board);
 
     public abstract Type getType();
 }
