+++ src/main/java/janggi/piece/Soldier.java
@@ -32,11 +32,6 @@ public final class Soldier extends Piece {
     }
 
     @Override
-    protected boolean isValidDestination(final Position end, final Board board) {
-        return !board.isPresentSameTeam(team, end);
-    }
-
-    @Override
     public Type getType() {
         return Type.SOLDIER;
     }
